move i18n package outside legacy
