+++ legacy/builder/additional_sketch_files_copier.go
@@ -32,7 +32,7 @@ package builder
 import (
 	"bytes"
 
-	"github.com/arduino/arduino-cli/legacy/builder/i18n"
+	"github.com/arduino/arduino-cli/arduino/i18n"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
 	"github.com/arduino/go-paths-helper"
 )
