+++ legacy/builder/read_file_and_store_in_context.go
@@ -30,7 +30,7 @@
 package builder
 
 import (
-	"github.com/arduino/arduino-cli/legacy/builder/i18n"
+	"github.com/arduino/arduino-cli/arduino/i18n"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
 	"github.com/arduino/go-paths-helper"
 )
